Update bindgen to 0.72 (#41)
The old bindgen 0.58 produced just a tiny bit of output and failed silently, the new version seems to work after tweaking a few things.

## TODO

- [x] Fix clippy
- [x] Fix tests

+++ nv-video-codec/src/encoder/error.rs
@@ -96,7 +96,7 @@ impl IntoNvEncResult for NVENCSTATUS {
                 Err(NvEncError::ResourceNotRegistered)
             },
             NVENCSTATUS::NV_ENC_ERR_RESOURCE_NOT_MAPPED => Err(NvEncError::ResourceNotMapped),
-            _ => Err(NvEncError::Unrecognized(self.0 as u32)),
+            _ => Err(NvEncError::Unrecognized(self.0)),
         }
     }
 }
